Исправление идеи и способа шифрования

+++ Encryptions/Models/IEncryption.cs
@@ -4,7 +4,7 @@ namespace TheoryOfInformation.lab1.Encryptions.Models
 {
     public interface IEncryption
     {
-        string BuildKey(ulong beginState, ushort length);
+        string BuildKey(ulong beginState, ulong length);
         byte[] Encrypte(byte[] file, byte[] key);
         BigInteger Encrypte(BigInteger text, BigInteger key);
     }
